refactor file positions
