+++ src/main/java/Employee_demo/EmployeeController.java
@@ -1,11 +1,9 @@
-package com.securian.creditcompass;
+package Employee_demo;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.util.Optional;
 
 @RestController
 public class EmployeeController {
