+++ src/main/java/Employee_demo/EmployeeRepository.java
@@ -1,6 +1,6 @@
-package com.securian.creditcompass;
+package Employee_demo;
 
+import Employee_demo.Employee;
 import com.securian.creditcompass.CrudRepository;
-import com.securian.creditcompass.Employee;
 
 public interface EmployeeRepository extends CrudRepository<Employee, Integer> {}
